vision-service/requirements.txt
fastapi
uvicorn
opencv-python-headless
pydantic
numpy
python-dotenv
redis
paddlepaddle
paddleocr
onnxruntime
ImageHash
Pillow
groq
